Create GitHub Actions workflow for periodic commits
Added a new GitHub Actions workflow that commits a readme.txt file to the 'keep-gh-actions-alive' branch every 50 days. This is to ensure that GitHub Actions do not get automatically disabled due to inactivity. The readme.txt file will contain a message indicating the purpose of the commit along with the last updated date and time.